backend/requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
psycopg2-binary>=2.9.9
# SQLite is built into Python

# HTTP client
httpx>=0.27.0

# Image processing
Pillow>=10.2.0

# Environment
python-dotenv>=1.0.0

# CORS
starlette>=0.35.0
